requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.3
Flask-Admin==1.6.1
python-dotenv==1.0.0
requests==2.31.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
celery==5.5.3
redis==5.0.1
Werkzeug==3.0.1